Use 'YEAR' part of a date to determine the leap year.
In order to calculate the last day of the month, we were using 'DAY'
part of a date to determine the leap year, which was wrong. We should
have used 'YEAR' part for the date for that caculation as proposed in
this change.

Original Patch by: Mehmet Emin KARAKAŞ <emin100@gmail.com>

Bumped the version of pgAgent to 4.1.
Created the upgrade sql scripts to allow upgrading from older version
(i.e. unpacked/3.4/4.0) to 4.1 (latest) version.

Reviewed by: Neel Patel